add books entity, integration with author

+++ src/authors/authors.service.ts
@@ -1,14 +1,21 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, BadRequestException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm'
 import { CreateAuthorDto } from './dto/create-author.dto';
 import { UpdateAuthorDto } from './dto/update-author.dto';
 import { MongoRepository } from 'typeorm';
 import { Author } from './entities/author.entity';
+import { Book } from 'src/books/entities/book.entity';
 import { ObjectID } from 'mongodb';
 
 @Injectable()
 export class AuthorsService {
+  // constructor(
+  //   @InjectRepository(Author)
+  //   private readonly authorRepository: MongoRepository<Author>
+  // ) {}
   constructor(
+    @InjectRepository(Book)
+    private readonly bookRepository: MongoRepository<Book>,
     @InjectRepository(Author)
     private readonly authorRepository: MongoRepository<Author>
   ) {}
@@ -29,11 +36,22 @@ export class AuthorsService {
   async update(id: string, updateAuthorDto: UpdateAuthorDto) {
     const _id = ObjectID(id);
     let author = await this.authorRepository.findOne({_id});
+    if(!author) {
+      throw new BadRequestException("can't find the author");
+    }
     let updated = Object.assign(author, updateAuthorDto);
     return await this.authorRepository.save(updated);
   }
 
   async remove(id: string) {
+    const books = await this.bookRepository.find({
+      where: { 
+        author: {$eq: id}, 
+      } 
+    });
+    if(books.length) {
+      throw new BadRequestException("At first remove book of that author");
+    }
     const _id = ObjectID(id);
     let deleted = await this.authorRepository.delete({_id});
     return deleted?.raw?.result;
